requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv

